add linter

+++ pkg/runtime.go
@@ -17,37 +17,51 @@ func StartTestSuite(ctx context.Context, suiteName string) error {
 	suiteName = suiteName + ".yaml"
 	suiteFilePath := filepath.Join(dir, suiteDirectory, suiteName)
 
+	dummyNetwork := true
+
 	cfg, err := LoadTestSuite(suiteFilePath)
 	if err != nil {
 		return err
 	}
 
 	// todo: spawn kurtosis gateway?
-
-	kurtosisCtx, err := GetKurtosisContext()
-	if err != nil {
-		return err
-	}
-
-	enclaveCtx, err := CreateEnclaveContext(ctx, kurtosisCtx)
-	namespace := fmt.Sprintf("kt-%s", enclaveCtx.GetEnclaveName())
-	if err != nil {
-		return err
-	}
-
-	defer func() {
-		err := kurtosisCtx.DestroyEnclave(ctx, enclaveCtx.GetEnclaveName())
+	var namespace string
+	if !dummyNetwork {
+		kurtosisCtx, err := GetKurtosisContext()
 		if err != nil {
-			log.Fatal(err)
+			return err
 		}
-	}()
 
-	StartNetwork(ctx, enclaveCtx, cfg.HarnessConfig)
+		enclaveCtx, err := CreateEnclaveContext(ctx, kurtosisCtx)
+		namespace = fmt.Sprintf("kt-%s", enclaveCtx.GetEnclaveName())
+		if err != nil {
+			return err
+		}
 
-	//namespace := "kt-attacknet-1696449504"
-	stopCh, err := CreateGrafanaClient(ctx, namespace, cfg.AttacknetConfig)
+		defer func() {
+			err := kurtosisCtx.DestroyEnclave(ctx, enclaveCtx.GetEnclaveName())
+			if err != nil {
+				log.Fatal(err)
+			}
+		}()
+
+		StartNetwork(ctx, enclaveCtx, cfg.HarnessConfig)
+	} else {
+		namespace = "kt-ethereum"
+	}
+
+	grafanaTunnel, err := CreateGrafanaClient(ctx, namespace, cfg.AttacknetConfig)
+	if err != nil {
+		return err
+	}
 	defer func() {
-		close(stopCh)
+		close(grafanaTunnel.PortForwardStopCh)
 	}()
-	return err
+
+	//ds, err := grafanaTunnel.Client.GetDatasource(ctx, 1)
+
+	//grafanaTunnel.Client.CreateAlertNotification()
+
+	//_ = ds
+	return nil
 }
